requirements.txt
Flask==1.1.1
future==0.17.1
gcloud==0.18.3
gcs-oauth2-boto-plugin==2.5
ggplot==0.11.5
google==2.0.2
google-api-core==1.14.0
google-apitools==0.5.30
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-core==1.0.3
google-cloud-pubsub==0.44.0
google-cloud-storage==1.17.0
google-cloud-translate==1.6.0
google-cloud-vision==0.38.0
google-reauth==0.1.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.12.3
gspread==3.1.0
grpcio==1.22.0
gsutil==4.41
httplib2==0.13.1
line-bot-sdk==1.12.1
MarkupSafe==1.1.1
mccabe==0.6.1
mock==2.0.0
monotonic==1.5
networkx==2.3
oauth2client==4.1.3
opencv-python==4.1.0.25
pandas==0.24.2
patsy==0.5.1
pbr==5.4.1
pickle-mixin==1.0.2
Pillow==6.0.0
protobuf==3.9.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pylint==2.3.1
requests==2.22.0